# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GlobeRise Backend API",
    "description": "DeFi and MLM platform backend API with 4-Wallet architecture, investment packages, referral system, and blockchain integration.",
    "version": "1.0.0",
    "contact": {
      "name": "GlobeRise Support",
      "email": "support@globerise.com"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "http://localhost:6969/api",
      "description": "Development server"
    },
    {
      "url": "https://api.globerise.com/api",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/transactions/earnings": {
      "get": {
        "tags": ["Transactions"],
        "summary": "Get earnings summary",
        "description": "Get earnings summary breakdown by type",
        "operationId": "getEarningsSummary",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Earnings summary retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "message": { "type": "string" },
                    "data": { "$ref": "#/components/schemas/EarningsSummary" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "EarningsSummary": {
        "type": "object",
        "properties": {
          "breakdown": {
            "type": "object",
            "additionalProperties": {
              "type": "object",
              "properties": {
                "total": { "type": "string" },
                "count": { "type": "integer" }
              }
            }
          },
          "totalEarnings": { "type": "string" }
        }
      },
...
    }
  }
}